pyproject.toml
[tool.poetry]
name = "nexustask-enterprise"
version = "10.4.2"
description = "Enterprise-grade, highly scalable, distributed task and project management system."
authors = ["NexusTask Core Team <core@nexustask.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "nexustask", from = "src"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Office/Business :: Scheduling",
    "Enterprise :: Scale :: Global"
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
alembic = "^1.11.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
click = "^8.1.0"
rich = "^13.0.0"
structlog = "^23.1.0"
opentelemetry-api = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
prometheus-client = "^0.17.0"
redis = "^4.6.0"
celery = "^5.3.0"
kombu = "^5.3.0"
graphql-core = "^3.2.3"
strawberry-graphql = "^0.200.0"
tenacity = "^8.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
kafka-python = "^2.0.2"
elasticsearch = "^8.8.0"
boto3 = "^1.28.0"
aiohttp = "^3.13.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.0"
flake8 = "^6.0.0"
tox = "^4.6.0"
locust = "^2.15.0"
sphinx = "^7.0.0"
mkdocs-material = "^9.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
